Refactor

+++ jobManager.js
@@ -1,15 +1,14 @@
 var fs = require('fs')
   , logger = require('./logger')
   , path = require('path')
-  , settings = require('./config')
   , jobsFolder = path.join(__dirname, 'jobs')
   , _ = require('lodash')
   , registry = {}
   , logger = require('./logger');
 
-function registerJobs(){
+function registerJobs() {
   logger.info('registering jobs');
-
+  
   var jobs = fs.readdirSync(jobsFolder);
 
   jobs.forEach(function(key){
@@ -20,21 +19,25 @@ function registerJobs(){
     var module = require(path.join(jobsFolder,key));
 
     registry[key] = {
-      name:module.name,
+      name: module.name,
       job: module,
       children: [],
       spawn: function(){
         logger.info('spawning child: ' + this.name);
-        var child = this.job.spawn();
-        child.started = new Date();
-        this.children.push(child);
+        var child = this.job.spawn(logger);
 
-        //bind child to handlers
-        process.on('exit', function() {
-          child.kill();
-        });
+        // If child is a forced process...
+        if(child != undefined && child.pid != undefined) {
+          child.started = new Date();
+          this.children.push(child);
+
+          //bind child to handlers
+          process.on('exit', function() {
+            child.kill();
+          });
+        }
       },
-      methods:module.methods
+      methods: module.methods
     };
 
   })
@@ -43,19 +46,13 @@ function registerJobs(){
 }
 
 var load = function(){
-  var eagerSpawn = settings.eagerSpawn;
-
   registerJobs();
 
   Object.keys(registry)
     .forEach(function(registrationName){
       var registration = registry[registrationName];
-
-      logger.info('checking eager spawn for ' + registration.name);
-      if(eagerSpawn === true || eagerSpawn == '*' || eagerSpawn.indexOf('*') > -1 || eagerSpawn.indexOf(registration.name) > -1)
-        registration.spawn();
-
-    })
+      registration.spawn();
+    });
 }
 
 module.exports = {
